+++ database/migrations/2023_11_08_155810_create_user_events_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('user_events', function (Blueprint $table) {
             $table->uuid('ue_id')->primary();
-            $table->integer('User_u_id')->index();
+            $table->unsignedBigInteger('User_u_id')->index();
             $table->uuid('Event_e_id')->index();
 
             $table->foreign('User_u_id')->references('id')->on('users');
